Recipe.md
to make this pizza, add warm water and yeast and mix it well. Then add sugar and let it sit for 5 min. Then add flour and salt and make a dough. Knead the dough for 15 min and apply oil
when it gets sticky. After kneading let it rest for 1 hour. after one hour , roll out the dough ina circle and add sauce and cheese. Bake it at 200 degree celcius for 15 to 20 min, or untill
the cheese is golden brown.